+++ ui/app/containers/Walk/index.jsx
@@ -3,10 +3,7 @@ import { Helmet } from 'react-helmet-async';
 import { useDispatch, useSelector } from 'react-redux';
 import { useInjectReducer, useInjectSaga } from 'redux-injectors';
 
-import { getHostToken } from '../../utils/host';
-
 import actions from './actions';
-import { resizeDimensions } from '../../../../config.json';
 import reducer from './reducer';
 import saga from './saga';
 import { selectFiles, selectPath } from './selectors';
@@ -17,7 +14,9 @@ import GenericList from '../../components/GenericList';
 import ListFile from './ListFile';
 import LoadingIndicator from '../../components/LoadingIndicator';
 import Menu from './Menu';
-import OrganizePreviews from '../../components/OrganizePreviews';
+import OrganizePreviews, {
+  DraggableThumb,
+} from '../../components/OrganizePreviews';
 
 const {
   addParentDirectoryNav,
@@ -25,8 +24,6 @@ const {
   parseHash,
   organizeByMedia,
 } = walkUtils;
-
-const CDN_HOST = getHostToken('cdn');
 
 function Walk({ location: { hash } }) {
   const dispatch = useDispatch();
@@ -81,31 +78,16 @@ function Walk({ location: { hash } }) {
       loading={loading}
       error={false}
     />,
+    hasImages && (
     <OrganizePreviews
       key="walk-OrganizePreviews"
       setItems={setItems}
       items={stateImages.map(item => ({
         ...item,
-        content: [
-          <span key={`label-${item.filename}`}>
-            <A
-              href={`${CDN_HOST}/${statePath}/${item.filename}`}
-              target="_blank"
-              title="View original in new tab"
-            >
-              {item.filename}
-            </A>
-          </span>,
-          <img
-            key={`thumbnail-${item.filename}`}
-            alt="No preview yet"
-            src={`${CDN_HOST}/${statePath}/${item.filename}`}
-            width={resizeDimensions.preview.width}
-            height={resizeDimensions.preview.height}
-          />,
-        ],
+          content: <DraggableThumb parentCdnFolder={statePath} item={item} />,
       }))}
-    />,
+      />
+    ),
   ];
 }